requirements.txt
# Dify Plugin Framework
dify_plugin>=0.2.0,<0.3.0

# Huawei Cloud OBS SDK
esdk-obs-python>=3.24.0